Update to truffle 5.1 (#21)
This PR is a prerequisite for https://github.com/gnosis/mock-contract/pull/20 in that it fixes the web3 API and unit tests to use a more recent truffle version. This way the other PR can only focus on updating the smart contract for solidity v0.6

The main changes are adjusting the unit test to use the web3 API (e.g. `contracts.methods`). This also means we no longer need the `ethereumjs-abi` dependency. I also updated the README.

### Test Plan

CI